refactor : 0912 fe code combine
