+++ client/.eslintrc.json
@@ -36,3 +36,4 @@
       "react/prop-types": "off"
     }
   }
+}
