--- client/src/Components/Button.js
@@ -1,14 +1,0 @@
-import { styled } from "styled-components";
-
-export const Button = styled.button`
-  display: flex;
-  width: 124px;
-  height: 61px;
-  padding-bottom: 5px;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-  flex-shrink: 0;
-  border-radius: 24px;
-  background: #ff6347;
-`;
