+++ client/src/index.css
@@ -56,7 +56,7 @@
   margin: 0px;
   padding: 0px;
   box-sizing: border-box;
-  font-family: 'intelone-mono-font-family-regular', 'NanumSquareNeo-Variable', 'Giants-Bold','Single Day', 'Ramche', 'DOSPilgiMedium', 'HakgyoansimGaeulsopungB',  sans-serif;
+  font-family: "HakgyoansimBunpilR", 'intelone-mono-font-family-regular', 'NanumSquareNeo-Variable', 'Giants-Bold','Single Day', 'Ramche', 'DOSPilgiMedium', 'HakgyoansimGaeulsopungB',  sans-serif;
 }
 
 html {
