+++ client/src/pages/UserModify.jsx
@@ -1,4 +1,5 @@
 import { styled } from "styled-components";
+import { React } from "react";
 
 export default function UserModify() {
   return (
