+++ client/src/pages/styles/SignUpBox.js
